requirements.txt
Django==1.8
unipath
python-dateutil
pytz
